bugfix: pr should be float, not int
Former-commit-id: 0a2f88f3b29053dfd54f7a7d4921fe05265aff87

+++ single_cell/workflows/singlecell_summary/scripts/plot_pcolormesh.py
@@ -174,7 +174,7 @@ class PlotPcolor(object):
 
             numreads = int(line[idxs['total_mapped_reads']])
 
-            cell_quality = int(line[idxs['probability_good']])
+            cell_quality = float(line[idxs['probability_good']])
 
             if self.cellcalls and cc not in self.cellcalls:
                 continue
